backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
pyodbc>=5.1.0
python-dotenv==1.0.1
httpx==0.27.2
gunicorn==22.0.0
